+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - API Scenario
   - Support scenario file as scope
   - Support test-proxy recording externalization with `--testProxyAssets` option
+  - Support `--randomSeed` option
 
 ## 12/01/2022 3.2.4